Move map, create getter methods, update javadoc
