+++ Pathfinding/src/main/java/me/nabdev/pathfinding/Pathfinder.java
@@ -5,6 +5,7 @@ import me.nabdev.pathfinding.algorithms.SearchAlgorithm;
 import me.nabdev.pathfinding.algorithms.SearchAlgorithm.SearchAlgorithmType;
 import me.nabdev.pathfinding.structures.Edge;
 import me.nabdev.pathfinding.structures.ImpossiblePathException;
+import me.nabdev.pathfinding.structures.Map;
 import me.nabdev.pathfinding.structures.Obstacle;
 import me.nabdev.pathfinding.structures.Path;
 import me.nabdev.pathfinding.structures.Vertex;
@@ -143,7 +144,7 @@ public class Pathfinder {
         map = new Map(obstacles, obstacleVertices, edges, clearance);
 
         for (Obstacle obs : obstacles) {
-            obs.initialize(map.pathVerticesStatic);
+            obs.initialize(map.getPathVerticesStatic());
         }
     }
 
@@ -410,10 +411,10 @@ public class Pathfinder {
      */
     public ArrayList<Edge> visualizeNeighbors() {
         ArrayList<Edge> list = new ArrayList<>();
-        if (map.neighbors != null)
-            list.addAll(map.neighbors);
-        if (map.neighborsStatic != null)
-            list.addAll(map.neighborsStatic);
+        if (map.getNeighbors() != null)
+            list.addAll(map.getNeighbors());
+        if (map.getNeighborsStatic() != null)
+            list.addAll(map.getNeighborsStatic());
         return list;
     }
 
@@ -427,8 +428,8 @@ public class Pathfinder {
      */
     public ArrayList<Vertex> visualizePathVertices() {
         ArrayList<Vertex> list = new ArrayList<>();
-        if (map.pathVertices != null)
-            list.addAll(map.pathVertices);
+        if (map.getPathVertices() != null)
+            list.addAll(map.getPathVertices());
         return list;
     }
 
@@ -466,7 +467,7 @@ public class Pathfinder {
      */
     public ArrayList<Vertex> visualizeInflatedVertices() {
         ArrayList<Vertex> list = new ArrayList<>();
-        list.addAll(map.pathVerticesStatic);
+        list.addAll(map.getPathVerticesStatic());
         return list;
     }
 }
